requirements.txt
Flask==2.1.0
openai==0.27.0
python-dotenv==0.19.1

